Use MUnit in FrameSerializerSuite
