+++ project/Dependencies.scala
@@ -1,6 +1,8 @@
 import sbt._
 
 object Dependencies {
+  private val munitVersion = "0.7.26"
+
   lazy val logbackClassic      = "ch.qos.logback"             %  "logback-classic"     % "1.2.3"
   lazy val twitterHPACK        = "com.twitter"                %  "hpack"               % "1.0.2"
   lazy val asyncHttpClient     = "org.asynchttpclient"        %  "async-http-client"   % "2.12.3"
@@ -9,5 +11,6 @@ object Dependencies {
   lazy val specs2              = "org.specs2"                 %% "specs2-core"         % "4.12.1"
   lazy val specs2Mock          = "org.specs2"                 %% "specs2-mock"         % specs2.revision
   lazy val specs2Scalacheck    = "org.specs2"                 %% "specs2-scalacheck"   % specs2.revision
-  lazy val munit               = "org.scalameta"              %% "munit"               % "0.7.26"
+  lazy val munit               = "org.scalameta"              %% "munit"               % munitVersion
+  lazy val scalacheckMunit     = "org.scalameta"              %% "munit-scalacheck"    % munitVersion
 }
